Format into string Function is finished. Also include the test file.
