+++ test-it/PrintFormatting_Numeric.via
@@ -2,62 +2,67 @@
 define (HelloWorld dv(.VirtualInstrument (
 Locals: c( 
 	  e(dv (.Int32 71) num1) 
-	  e(dv (.Int32 72) num2) 
-	  e(dv (.Int32 73) num3) 
+	  e(dv (.Int32 64) num2) 
+	  e(dv (.Int32 -73) num3) 
 	  e(dv (.Int32 74) num4) 
      	  e(dv (.Double 162.34630) db1)
-     	  e(dv (.Double 0.0199360) db6)
-	  e(dv (.Double 9.96)db2)
+  	  e(dv (.Double 9.96)db2)
 	  e(dv (.Double 999.9996)db3)
 	  e(dv (.Double 13999.573)db4)
  	  e(dv (.Double 7.573)db5)
+     	  e(dv (.Double 0.0199360) db6)
      	  e(dv (.Double -0.000197960) db7)
+     	  e(dv (.Double 0.000000000000000123) db8)
+     	  e(dv (.Double 893545675713456343412) db9) 
 	  e(dv (.Double 0) db0)
 	  e(dv(.String "abcde") s1) 
 	  e(dv(.String "Hello") s3) 
 	  e(dv(.String "你好") s4) 
 	) 
 	clump (
- 	  Printf("Test width sepcifer-------------------------\n")
- 	  Printf("Format String:%%47s:<%s>     %%00011d:-%d-\n" s1   num1)
- 	  Printf("<%47s>     -%00011d-\n" s1   num1)
- 	  Printf("Test variable order-------------------------\n")
-	  Printf("%%4$d %%3$d  %%2$d  %%1$d => %d %d %d %d \n" num1 num2 num3 num4   )
-	  Printf("%4$d %3$d %2$d %1$d \n" num1 num2 num3 num4   )
- 	  Printf("Test variable order-------------------------\n")
-	  Printf("%%d %%4$d %%d %%1$d %%s=>%d %d %s %d %d \n" num1 num2 s1 num3 num4  )
-	  Printf("%d %4$d %d %1$d %s \n" num1 num2 s1 num3 num4   )
- 	
- 	  Printf("Test Float %f-------------------------\n" db1)
+	  Printf("Test Integer %d-------------------------\n")
+	  Printf("%%020d %1$d=> %020d\n" num1)
+	  Printf("%%020d %1$d=> %020d\n" db1)
+	  Printf("%%+020x %1$d=> %020x\n" num4)
+	  Printf("%%#020x %1$d=> %#020x\n" db3)
+
+	  Printf("\nTest Binary %d-------------------------\n")
+	  Printf("%%b %1$d=> %b\n" num1)
+	  Printf("%%b %1$d=> %b\n" num2)
+	  Printf("%%b %1$d=> %b\n" num3)
+ 	  Printf("\nTest Float %f -------------------------\n" db1)
 	  Printf("%%010_5f => *%010_5f*\n" db1  )
 	  Printf("%%+_5f   => *%+_5f*\n" db1  )
 	  Printf("%%10_2f  => *%10_2f*\n" db1  )
-	  Printf("%%10.5f  => *%10.5f*\n" db1  )
- 	  Printf("%% 10.5f  => *% 10.5f*\n" db1  )
+	  Printf("%%+10.5f => *%+10.5f*\n" db1  )
+ 	  Printf("%%10.5f  => *%10.5f*\n" db1  )
 	  Printf("%%#10.6f => *%#10.6f*\n" db1  )
- 	  Printf("Test Float %f-------------------------\n" db2)
+ 	  Printf("\nTest Float %f -------------------------\n" db2)
 	  Printf("%%010_1f => *%010_1f*\n" db2  )
-	  Printf("%%+_5f   => *%+_5f*\n" db2  )
+	  Printf("%%+_0f   => *%+_0f*\n" db2  )
 	  Printf("%%10_2f  => *%10_2f*\n" db2 )
 	  Printf("%%-10.1f => *%-10.1f*\n" db2  )
 	  Printf("%%#10.6f => *%#10.6f*\n" db2  )
-	  Printf("Test Float %f-------------------------\n" db6)
+
+	  Printf("\nTest Float %f-------------------------\n" db6)
 	  Printf("%%010_1f => *%010_1f*\n" db6  )
 	  Printf("%%+_5f   => *%+_5f*\n" db6 )
 	  Printf("%%10_2f  => *%10_2f*\n" db6 )
 	  Printf("%%-10.1f => *%-10.1f*\n" db6  )
+	  Printf("%%010_1f => *%010_1f*\n" db6  )
 	  Printf("%%#10.6f => *%#10.6f*\n" db6  )
- 	  Printf("Test Exp %e-------------------------\n" db3)
+ 	  Printf("\nTest Exp %e-------------------------\n" db3)
 	  Printf("%%+10.5e => *%+10.5e*\n" db3  )
 	  Printf("%%010_5e  => *%010_5e*\n" db3  )
 	  Printf("%%#10.6e  => *%#10.6e*\n" db3  )
 	  Printf("%%+^010_3e=> *%+^010_3e*\n" db3  )
-	  Printf("Test Exp %e-------------------------\n" db4)
+	  Printf("%%^10.5e  => *%^10.5e*\n" db3  )
+	  Printf("\nTest Exp %e-------------------------\n" db4)
 	  Printf("%%+10.5e => *%+10.5e*\n" db4  )
 	  Printf("%%010_5e  => *%010_5e*\n" db4  )
 	  Printf("%%^10.6e  => *%^10.6e*\n" db4  )
 	  Printf("%%+^010_3e=> *%+^010_3e*\n" db4  )
- 	  Printf("Test Exp %e-------------------------\n" db5)
+ 	  Printf("\nTest Exp %e-------------------------\n" db5)
 	  Printf("%%+10.1e => *%+10.1e*\n" db5  )
 	  Printf("%%010_5e  => *%010_5e*\n" db5  )
 	  Printf("%%^10.6e  => *%^10.6e*\n" db5 )
@@ -67,7 +72,7 @@ Locals: c(
 	  Printf("%%#010_5e  => *%#010_5e*\n" db7  )
 	  Printf("%%^10.6e   => *%^10.6e*\n" db7 )
 	  Printf("%%+^010_3e => *%+^010_3e*\n" db7  )
- 	  Printf("Test Zero-------------------------\n" db0)
+ 	  Printf("\nTest Zero-------------------------\n" db0)
 	  Printf("%%+10.5e   => *%+10.5e*\n" db0  )
 	  Printf("%%#0_5e  => *%#0_5e*\n" db0  )
 	  Printf("%%^10.6e   => *%^10.6e*\n" db0)
@@ -77,21 +82,30 @@ Locals: c(
 	  Printf("%%10_2f  => *%10_2f*\n" db0 )
 	  Printf("%%-10.1f => *%-10.1f*\n" db0  )
 	  Printf("%%#10.6f => *%#10.6f*\n" db0  )
-	  Printf("Test int to %%e & %%f-------------------------\n")
+	  Printf("\nTest int to %%e & %%f-------------------------\n")
 	  Printf("%%+10.2f   => *%+10.2f*\n" num1 )
-	  Printf("%%#0_5f  => *%#0_5f*\n" num2  )
+	  Printf("%%#0_5f    => *%#0_5f*\n" num2  )
 	  Printf("%%^10.6e   => *%^10.6e*\n" num3)
 	  Printf("%%+^010_3f => *%+^010_3e*\n" num4  )
- 	  Printf("Test %%g-------------------------\n" db0)
-	  Printf("%%+10.5g   => *%+10.5g*\n" db1 )
-	  Printf("%%#0_5g  => *%#0_5g*\n" db1  )
-	  Printf("%%^10.6g   => *%^10.6g*\n" db2)
-	  Printf("%%+^010_3g => *%+^010_3g*\n" db3  )
-  	  Printf("%%010_1g => *%010_1g*\n" db4  )
-	  Printf("%%#+_5g   => *%#+_5g*\n" db5  )
-	  Printf("%%10_2g  => *%10_2g*\n" db6 )
-	  Printf("%%-10.1g => *%-10.1g*\n" db7  )
-	  Printf("%%#10.6g => *%#10.6g*\n" db0  )
+ 	  Printf("\nTest %%g-------------------------\n")
+	  Printf("%%+10.5g  %1$f => *%+10.5g*\n" db1 )
+	  Printf("%%#0_5g   %1$f => *%#0_5g*\n" db1  )
+	  Printf("%%^10.6g  %1$f => *%^10.6g*\n" db2)
+	  Printf("%%+^010_3g %1$f=> *%+^010_3g*\n" db3  )
+  	  Printf("%%010_1g  %1$f => *%010_1g*\n" db4  )
+	  Printf("%%+_5g    %1$f => *%+_5g*\n" db5  )
+	  Printf("%%10_2g   %1$f => *%10_2g*\n" db6 )
+	  Printf("%%-10.1g  %1$f => *%-10.1g*\n" db7  )
+	  Printf("%%10.6g   %1$e => *%10.6g*\n" db0  )
+	  Printf("%%10.5g   %1$f => *%10.5g*\n" db8  )
+  	  Printf("\nTest %%p-------------------------\n")
+	  Printf("%%p   %1$f     => *%p*\n" num1 )
+	  Printf("%%+10.5p  %1$f  => *%+10.5p*\n" db1 )
+	  Printf("%%#0_5p   %1$f => *%#0_5p*\n" db8  )
+	  Printf("%%0_5p    %1$f  => *%0_5p*\n" db8  )
+	  Printf("%%p  %1$f => *%p*\n" db9)
+	  Printf("%%^10.2p  %1$f => *%^10.6p*\n" db9)
+	 
  	)
 ) ) )
 enqueue (HelloWorld)
