+++ test-it/PrintFormatting_VariableOrder.via
@@ -1,0 +1,29 @@
+ // test different format code
+define (HelloWorld dv(.VirtualInstrument (
+Locals: c( 
+	  e(dv (.Int32 71) num1) 
+	  e(dv (.Int32 72) num2) 
+	  e(dv (.Int32 73) num3) 
+	  e(dv (.Int32 74) num4) 
+     	  e(dv (.Double 162.34630) num5)
+     	  e(dv (.Double 89.54630) num6)
+	  e(dv(.String "abcde") s1) 
+	  e(dv(.String "") s2) 
+	  e(dv(.String "你好") s4) 
+	) 
+	clump (
+ 	  Printf("Test width sepcifer-------------------------\n")
+ 	  Printf("Format String:%%47s:<%s>     %%00011d:-%d-\n" s1   num1)
+ 	  Printf("<%47s>     -%00011d-\n" s1   num1)
+ 	  Printf("Test variable order-------------------------\n")
+	  Printf("%%4$d %%3$d  %%2$d  %%1$d => %d %d %d %d \n" num1 num2 num3 num4   )
+	  Printf("%4$d %3$d %2$d %1$d \n" num1 num2 num3 num4   )
+ 	  Printf("Test variable order-------------------------\n")
+	  Printf("%%d %%4$d %%d %%1$d %%s => %d %d %s %d %d \n" num1 num2 s4 num3 num4  )
+	  Printf("%d %4$d %% %hisiserror %d %mon %1$d %s\n" num1 num2 s4 num3 num4)
+
+ 	  Printf("%%d %%d %%5$d %%4$d %%3$s=>%d %d %s %d %d \n" num1 num2 s1 num5 num6 )
+	  Printf("%d %d %5$d %+#.12mm %4$d %12$[] %3$s\n" num1 num2 s1 num5 num6)
+ 	)
+) ) )
+enqueue (HelloWorld)
