+++ test-it/results/PrintFormatting_Numeric.vtr
@@ -1,0 +1,94 @@
+Test Integer 0-------------------------
+%020d 71=> 00000000000000000071
+%020d 162=> 00000000000000000162
+%+020x 74=> 0000000000000000004a
+%#020x 1000=> 0x0000000000000003e8
+
+Test Binary 0-------------------------
+%b 71=> 1000111
+%b 64=> 1000000
+%b -73=> 11111111111111111111111110110111
+
+Test Float 162.346300 -------------------------
+%010_5f => *0000162.35*
+%+_5f   => *+162.35*
+%10_2f  => *       160*
+%+10.5f => *+162.34630*
+%10.5f  => * 162.34630*
+%#10.6f => *  162.3463*
+
+Test Float 9.960000 -------------------------
+%010_1f => *0000000010*
+%+_0f   => *+00*
+%10_2f  => *        10*
+%-10.1f => *10.0      *
+%#10.6f => *      9.96*
+
+Test Float 0.019936-------------------------
+%010_1f => *0000000.02*
+%+_5f   => *+0.019936*
+%10_2f  => *     0.020*
+%-10.1f => *0.0       *
+%010_1f => *0000000.02*
+%#10.6f => *  0.019936*
+
+Test Exp 9.999996E+2-------------------------
+%+10.5e => *+1.00000E+3*
+%010_5e  => *01.0000E+3*
+%#10.6e  => *9.999996E+2*
+%+^010_3e=> *+001.00E+3*
+%^10.5e  => *1.00000E+3*
+
+Test Exp 1.399957E+4-------------------------
+%+10.5e => *+1.39996E+4*
+%010_5e  => *01.4000E+4*
+%^10.6e  => *13.99957E+3*
+%+^010_3e=> *+0014.0E+3*
+
+Test Exp 7.573000E+0-------------------------
+%+10.1e => *   +7.6E+0*
+%010_5e  => *07.5730E+0*
+%^10.6e  => *7.573000E+0*
+%+^010_3e=> *+007.57E+0*
+Test Exp -1.979600E-4-------------------------
+%+10.5e   => *-1.97960E-4*
+%#010_5e  => *-1.9796E-4*
+%^10.6e   => *-197.9600E-6*
+%+^010_3e => *-000198E-6*
+
+Test Zero-------------------------
+%+10.5e   => *+0.00000E+0*
+%#0_5e  => *E+0*
+%^10.6e   => *0.000000E+0*
+%+^010_3e => *+000.00E+0*
+%010_1f => *0000000000*
+%#+_5f   => *+*
+%10_2f  => *       0.0*
+%-10.1f => *0.0       *
+%#10.6f => *          *
+
+Test int to %e & %f-------------------------
+%+10.2f   => *     +0.00*
+%#0_5f    => *-nan*
+%^10.6e   => *1.591497E-312*
+%+^010_3f => *-27.8E+171*
+
+Test %g-------------------------
+%+10.5g  162.346300 => **
+%#0_5g   162.346300 => **
+%^10.6g  9.960000 => **
+%+^010_3g 999.999600=> **
+%010_1g  13999.573000 => **
+%+_5g    7.573000 => **
+%10_2g   0.019936 => **
+%-10.1g  -0.000198 => **
+%10.6g   0.000000E+0 => **
+%10.5g   0.000000 => **
+
+Test %p-------------------------
+%p   0.000000     => **
+%+10.5p  162.346300  => **
+%#0_5p   0.000000 => **
+%0_5p    0.000000  => **
+%p  893545675713456373760.000000 => **
+%^10.2p  893545675713456373760.000000 => **
