+++ test-it/results/PrintFormatting_VariableOrder.vtr
@@ -1,0 +1,11 @@
+Test width sepcifer-------------------------
+Format String:%47s:<abcde>     %00011d:-71-
+<                                          abcde>     -00000000071-
+Test variable order-------------------------
+%4$d %3$d  %2$d  %1$d => 71 72 73 74 
+74 73 72 71 
+Test variable order-------------------------
+%d %4$d %d %1$d %s => 71 72 你好 73 74 
+71 73 % %isiserror 73 74on 71 ***TODO pointer type
+%d %d %5$d %4$d %3$s=>71 72 abcde 162 90 
+71 72 90 +000000000000m 162 0] 162.346
